--- src/Services/Catalog/Catalog.API/Infrastructure/IntegrationEventMigrations/20170322145434_IntegrationEventInitial.cs
@@ -1,32 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore.Migrations;
-
-namespace Catalog.API.Migrations
-{
-    public partial class IntegrationEventInitial : Migration
-    {
-        protected override void Up(MigrationBuilder migrationBuilder)
-        {
-            migrationBuilder.CreateTable(
-                name: "IntegrationEventLog",
-                columns: table => new
-                {
-                    EventId = table.Column<Guid>(nullable: false),
-                    Content = table.Column<string>(nullable: false),
-                    CreationTime = table.Column<DateTime>(nullable: false),
-                    EventTypeName = table.Column<string>(nullable: false),
-                    State = table.Column<int>(nullable: false),
-                    TimesSent = table.Column<int>(nullable: false)
-                },
-                constraints: table =>
-                {
-                    table.PrimaryKey("PK_IntegrationEventLog", x => x.EventId);
-                });
-        }
-
-        protected override void Down(MigrationBuilder migrationBuilder)
-        {
-            migrationBuilder.DropTable(
-                name: "IntegrationEventLog");
-        }
-    }
-}
