--- src/Services/Catalog/Catalog.API/Infrastructure/IntegrationEventMigrations/20190507184807_AddTransactionId.cs
@@ -1,22 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore.Migrations;
-
-namespace Catalog.API.Infrastructure.IntegrationEventMigrations
-{
-    public partial class AddTransactionId : Migration
-    {
-        protected override void Up(MigrationBuilder migrationBuilder)
-        {
-            migrationBuilder.AddColumn<string>(
-                name: "TransactionId",
-                table: "IntegrationEventLog",
-                nullable: true);
-        }
-
-        protected override void Down(MigrationBuilder migrationBuilder)
-        {
-            migrationBuilder.DropColumn(
-                name: "TransactionId",
-                table: "IntegrationEventLog");
-        }
-    }
-}
