--- src/Services/Catalog/Catalog.API/Infrastructure/IntegrationEventMigrations/IntegrationEventLogContextDesignTimeFactory.cs
@@ -1,14 +1,0 @@
-﻿namespace Catalog.API.Infrastructure.IntegrationEventMigrations
-{
-    public class IntegrationEventLogContextDesignTimeFactory : IDesignTimeDbContextFactory<IntegrationEventLogContext>
-    {
-        public IntegrationEventLogContext CreateDbContext(string[] args)
-        {
-            var optionsBuilder = new DbContextOptionsBuilder<IntegrationEventLogContext>();
-
-            optionsBuilder.UseSqlServer(".", options => options.MigrationsAssembly(GetType().Assembly.GetName().Name));
-
-            return new IntegrationEventLogContext(optionsBuilder.Options);
-        }
-    }
-}
